SVG labels (#29)
### [User Story
181787](https://dev.azure.com/EquinorASA/Spine/_workitems/edit/181787):
NOAKADEXPI some labels are not present in output svg

We were missing labels on the `IM005B`-symbols. This has been fixed now
by mapping the correct text to the correct label indexes. This has only
been done for `IM005B`-symbols for now, since these were the symbols
missing labels.

We want to do the same for all symbols, but it's going to take some work
to map out what text is to be placed on which label index, since this is
different for every symbol (this information can be found in the symbol
excel file). I have mad a new user story for this.

I have not done any work on the styling, as I think it will be easier
when we have mapped the label indexes for all symbols. I only added a
conditional value for `x`, as for default texts it needed to be `x -
70`, and for the new symbol it had to be only `x`. We can probably fix
the styling further at a later point, for example moving the texts
closer to each other inside the symbol.

Here is an image showing how the labels look now:

![image](https://github.com/user-attachments/assets/bb2ef6df-adeb-4e9d-90b1-7d12c673772f)