Selecting piping lines as boundaries (#26)
It is now possible to select the piping lines as boundaries. 

### Changes in xslt
It was difficult to style the symbols as they were wrapped within a use
and symbol tag. Since we have access to the svg code itself, wrapping
the code within symbol was not necessary. An additional xslt template
was necessary to match the new structure of the symbols properly.

### Changes in rml mappings
A new rml map is added to make the connectors between the piping
components imf:partOf the piping network segment.
This was necessary for the reasoning to work as intended since piping
network segments inherit the boundary from its child elements. Hence,
when selecting a pipe line (the connectors) as a boundary, this should
also be inherited to the piping network segment,.

### Changes in javascript 
- Added event listener for the piping nodes. 
- When a piping node is selected, they are marked red. This red markup
is a new element of class 'commissioningHighlight'. To be able to
deselect the boundary it was necessary to add an event listener to these
elements as well.
- New helper methods for updating the color of the pipes, and for
querying RDFox.
- Some refactoring

### UI 
1) When selecting a pipe as a boundary the pipe gets a red highlight. 
2) When the commissioning package is calculated, and if the pipe is
adjacent to a node that is within the commissioning package, the pipe
changes color to yellow. This is to indicate where the commissioning
package stops, _it goes up until the adjacent equipment, but not
including._ This is to make it possible to place a boundary on the sides
of a piping component.
3) If the commissioning package is "deselected" (removing the internal
node, yellow highlighting disappears), then the boundary pipes
previously within the commissioning package will be marked as red.

### Other changes
- Deleted some unused code
- Changed the name of some variables in the datalog. 
- Changes to styling in CSS