Delete unused mappings and renaming (#33)
## Clean up

### datalog
Some of the datalog rules was outdated, hence deleted. 

### doc
Deleted boundary.md and the corresponding images used in that document
as they are based on an old DEXPI example.

### rdfox
Deleted the rdfox script used on the old DEXPI example. I also renamed
the noaka.rdfox script to dexpi.rdfox.

### rml
Deleted the rml folder. This folder contained all rml mappings used on
the old DEXPI example.

### noaka_rml 
- Renamed this folder to rml_mappings 
- There is something weird in the git diff, it says that the
attrbutes-map.rml.ttl is deleted from the noaka_rml folder, however the
attributes map is still present in the renamed folder rml_mappings!
- Removed pytesting.py -> pushed by accident

### www 
- Deleted the old dexpi xslt mappings named dexpiOld.xslt
- Renamed dexpisvgNoaka.xslt to dexpisvg.xslt 

## Documentaton
I did not update any of the documentation on how to run the demo. This
is because we have a user story on removing the manual steps, this will
be prioritized in the upcoming sprints. Consequently, any updates to the
documentation would be rendered obsolete in the near future, making the
effort unnecessary.