Trying out docker for demo script (#39)
# Aim
Easier to run demo

# Method
Creating a docker-based setup that runs all scripts

# Consequences
Relative path to NOAKADEXPI repo had to be renamed in dexpisvg.xslt so
the path works in the docker image. This means running the
transformation locally does not work anymore.

dexpi.html is now hardcoded to have 14 lines before the svg. 

Rdfox license must now have a hardcoded position in the repo

Xslt transformation uses the unix tool xsltproc in stead of C#, since
faster and easier in docker

xslt transformation for rotation refactored to be calculated in C#

The result of the rml transformation is stored in the same location as
before (pandid.trig), while the generated svg is stored in
"www/index.html".

# User story: 
https://dev.azure.com/EquinorASA/Spine/_workitems/edit/190800

---------

Co-authored-by: Embla Øye <118742425+eoye@users.noreply.github.com>