Add graphical format example file (#48)
###
[AB#203007](https://dev.azure.com/EquinorASA/bb9bd8cb-74f7-4ffa-b0cb-60eff0a0be58/_workitems/edit/203007)

## Aim of the PR
A format for data to the frontend needs to be decided on. This format
should include everything needed to render the P&ID in the frontend.

## Implementation 
This PR includes a markdown file outlining what information the frontend
needs in order to display graphical P&ID data.

## Type of change
- [ ] Bug fix 
- [ ] New feature 
- [ ] Breaking change 
- [ ] This change requires a documentation update

## How Has This Been Tested?
- Tested that the JSON structure is legal by creating a JSON file in
WebStorm.
- Checked that it's possible to apply styled-component styles received
from a JSON file.