Symbol Translator (#47)
###
[AB#188605](https://dev.azure.com/EquinorASA/bb9bd8cb-74f7-4ffa-b0cb-60eff0a0be58/_workitems/edit/188605)

## Aim of the PR
Make the mapping from symbol name to label attributes available as JSON
for the frontend.
The assumption/decision is that the symbol-library is read by the
frontend, and not during translation on reception by backend


## Implementation 
Created a minimal asp.net web api with a single endpoint that takes a
DEXPI symbol ID as its only argument and returns a json structure like
this
```
{
  "id" : "ND0028",
  "description" : "Flow in nozzle",
  "labelAttributeA" : "<ObjectDisplayName>",
  "labelAttributeB" : "",
  "labelAttributeC" : "",
  "labelAttributeD" : "",
  "labelAttributeE" : ""
}
```
Also a dockerized version is made, which is now part of the docker
compose setup. Test the project by starting docker compose as usual,
described in docker/README.md


## Type of change
- [ ] Bug fix 
- [x] New feature 
- [ ] Breaking change 
- [ ] This change requires a documentation update

If the changes impact any dependent services then provide details.

## How Has This Been Tested?
There are github actions that test the labels A, B and C
The docker setup has been tried

## Open Questions
I am not sure if this is all data that is needed for the frontend, or if
the format is suitable for consumption by the frontend.