Story 201301: Preload SVGs to make P&ID rendering faster (#51)
## Aim of the PR
This PR fixes
[AB#201301](https://dev.azure.com/EquinorASA/bb9bd8cb-74f7-4ffa-b0cb-60eff0a0be58/_workitems/edit/201301)

When loading a P&ID in the frontend, some users experience very long
load times due to nested fetches (parent component fetches, then child
component, then grandchild component etc).

## Implementation 
When initially loading the P&ID, all SVGs are fetched simultaneously and
stored in a map from ComponentName to SVG element. This way we don't
need to perform a fetch in each element.

## Type of change
- [ ] Bug fix 
- [X] New feature 
- [ ] Breaking change 
- [ ] This change requires a documentation update

## How Has This Been Tested?
Tested load times on two different computers.

## Additional Changes
- Nozzle labels were placed far away from the nozzles. Label placement
is changed to be a bit better for nozzles. Rotation still not
implemented.
- Nozzle with ID Nozzle-7 lacks position and text information, and was
rendered in the upper left corner. Added a check for nozzle ID, and
skipped rendering for Nozzle-7.
- PipingComponent now uses SvgElement component for rendering.