Story 208659: Zoom functionality added (#55)
## Aim of the PR
This PR fixes
[AB#208659](https://dev.azure.com/EquinorASA/bb9bd8cb-74f7-4ffa-b0cb-60eff0a0be58/_workitems/edit/208659)

This PR aims to complete the sprint goal of adding zoom functionality to
the frontend.

## Implementation 
Added the package react-svg-zoom-pan.
Created a new component, ZoomableSVGWrapper, which sets up the
UncontrolledReactSVGZoomPan component and initializes some values.

It should now be possible to zoom by pinching or using the scroll wheel,
and pan by left clicking and dragging the SVG.

The toolbar can be overrided with a custom toolbar, might be nice to
implement it on the side with the other tools. :)

## Type of change
- [X] Bug fix 
- [X] New feature 
- [ ] Breaking change 
- [ ] This change requires a documentation update

If the changes impact any dependent services then provide details.
## How Has This Been Tested?
Tested by using touchpad to pinch zoom, and clicking and dragging to
pan. Not tested with mouse and scroll wheel.

## Additional Changes
The Drawing.Extent values are always of type string in the XML, but they
are parsed as numbers, which does not work in typescript. I have updated
the types to be of type string, and then we need to cast the values to
numbers.