Delete commissioning packages (#53)
[Pull requests practices for the SSI
team](https://github.com/equinor/ssi-infrastructure/blob/main/docs/pr_practices.md)

## Aim of the PR
This PR fixes
[AB#208957](https://dev.azure.com/EquinorASA/bb9bd8cb-74f7-4ffa-b0cb-60eff0a0be58/_workitems/edit/208957).

The goal of this PR is to add functionality for deleting commissioning
packages.

## Implementation 
I added a new sidebar item with a delete icon. When the user clicks the
button, a dialog opens. The dialog shows all commissioning packages, and
the user must choose which packages they want to delete. The active
package will be automatically checked, as it is likely the user wants to
delete the package that is currently active. When you delete the last
commissioning package, a new initial commissioning package will be
created.

## Type of change
- [ ] Bug fix 
- [x] New feature 
- [ ] Breaking change 
- [ ] This change requires a documentation update

## How Has This Been Tested?
This has been tested by running the demo and seeing that it is possible
to delete commissioning packages. See screenshots:

![image](https://github.com/user-attachments/assets/499cbf20-ba0c-4ff3-90f8-90670b169395)

![image](https://github.com/user-attachments/assets/0b9af7d5-14a8-4944-bf5a-953bb36be40a)

![image](https://github.com/user-attachments/assets/488464d0-4bb8-446c-8142-6def75d0aefa)