🌳 Stub endpoints for backend client 🌳 (#63)
## Aim of the PR
This PR fixes
[AB#209162](https://dev.azure.com/EquinorASA/bb9bd8cb-74f7-4ffa-b0cb-60eff0a0be58/_workitems/edit/209162)

Stubbed the endpoints that is needed for the dexpi demo. 

## NB
A lot of files changed since a new project is added! The main files are
the following:
- The contents of Backend/Model 
- Program.cs in Backend

Also did `dotnet format` which formatted the files in the other projects
in the Boundaries solution.

## Implementation 
Created a new web project in the Boundaries solution and created minimal
api's for stubbing the endpoints.
Also made made a proposal for two new DTO's: Node and
CommissioningPackage.
The endpoints cover the following operations:
- adding node as internal or boundary
- removing node as internal or boundary
- getting adjacent nodes of a node
- updating commissioning package color, name, and id.
- adding commissioning package
- deleting commissioning package
- getting a commissioning package

It is not created a basic endpoint for querying the triplestore, this is
intentional because the existing endpoints cover all necessary actions.
Also, exposing a generic endpoint might make the frontend harder to
maintain. If more specific endpoints is required then the backend should
create these.

## Type of change
- [ ] Bug fix 
- [x] New feature 
- [ ] Breaking change 
- [ ] This change requires a documentation update