Persist commissioning packages on refresh (#64)
[Pull requests practices for the SSI
team](https://github.com/equinor/ssi-infrastructure/blob/main/docs/pr_practices.md)

## Aim of the PR
This PR fixes
[AB#208958](https://dev.azure.com/EquinorASA/bb9bd8cb-74f7-4ffa-b0cb-60eff0a0be58/_workitems/edit/208958).

The goal of this PR is to make sure the commissioning packages are
persisted and not deleted on refresh of the page.

## Implementation 
Several changes have been made to achieve this:
- Instead of cleaning the triplestore and list of commissioning packages
on mount, we get the packages from the triplestore.
- When the app is started and an initial commissioning package is
created, the color and name is saved to the triplestore.
- When a user creates a new commissioning package, the color and name is
saved to the triplestore.
- Instead of getting just the node Ids from the triplestore, we now get
node Id, name and color.
- Added a new "getAllCommissioningPackages" that is called when the page
is refreshed. This method gets all commissioning packages, including the
name and color.
- Add `SelectedInternal` as type a node can have. This is so that we can
save which node was clicked as the internal boundary.

## Type of change
- [ ] Bug fix 
- [x] New feature 
- [ ] Breaking change 
- [ ] This change requires a documentation update

## How Has This Been Tested?
This has been tested by running the demo and testing the UI.

## Additional Changes
I deleted some unused code from `Triplestore.ts` to clean up the file.
Let me know if the code that has been removed should be kept there!