Add Undo and Redo functionality ↩️ (#66)
## Aim of the PR
This PR fixes
[AB#208953](https://dev.azure.com/EquinorASA/bb9bd8cb-74f7-4ffa-b0cb-60eff0a0be58/_workitems/edit/208953)

This PR implements undo and redo functionality to the DEXPI demo. This
allows the user to easily take back wrong actions.

## Implementation 

- The `Action` type has been created. This type contains the tool used,
and the ID of the affected node. The action history is an array of
`Action`.
- A context `ActionContext` has been created to keep track of the
current action.
- A new hook, `useUndoRedo` has been created. This hook exposes the
functions handleUndo and handleRedo. It also exposes the history state,
which is not used at the moment.
- `EditorSidebar` has been updated with undo and redo buttons.
- All clickable components now have the action context and pass this to
the handler functions.
- The handler functions have been refactored to update the action state.

## Type of change
- [ ] Bug fix 
- [X] New feature 
- [ ] Breaking change 
- [ ] This change requires a documentation update

If the changes impact any dependent services then provide details.
## How Has This Been Tested?
Please describe the tests that you ran to verify that your changes works
as expected.

## Additional Changes
- `HandlerFunctionHelper.tsx` has been renamed to
`CommissioningPackageHandler.tsx` to better reflect the role of the
functions in this file.
- This file has also been refactored for readability.

---------

Co-authored-by: Embla Øye <embla.oye@bouvet.no>