Use API endpoints in frontend (#80)
## Aim of the PR
This PR fixes
[AB#218921](https://dev.azure.com/EquinorASA/bb9bd8cb-74f7-4ffa-b0cb-60eff0a0be58/_workitems/edit/218921)

After creating an API for CRUD-operations against the triplestore, we
wish to use this API in the frontend instead of directly communicating
with the triplestore.

## Implementation 
**NOTE: The table is now broken, since those requests are handled by the
frontend. Will not be fixed until endpoints are in place for tables**


- **Api.ts**: This files handles request setup and sends API requests.
Right now, there is no ability to rename/change color of a commissioning
package, so this endpoint sits unfinished.
- **Triplestore.ts**: This file will be removed when all functionality
is covered by the API. For now, the unused functions are removed from
this file.
- **CommissioningPackageHandler.tsx**: A lot of the logic previously
contained in this file is now handled by the API.

## Type of change
- [ ] Bug fix 
- [X] New feature 
- [ ] Breaking change 
- [ ] This change requires a documentation update

## How Has This Been Tested?
Tested in demo that creation/deletion of package and selection of
boundary and internal in multiple commissioning packages works as
expected.
Tested the endpoints in Swagger to verify that the changes are updated. 
Tested in RdFox to see if correct nodes are marked as boundary/internal.

## Additional Changes
Some API fixes, docker setup fixes and datalog improvements needed to be
made for this PR to work as well.
Port 5000 is stupidly used by Apple AirPlay, causing issues when running
the dockerfiles on mac. Port is changed to 6000 for this reason.

Updates to Datalog rule, the following rule is added: 
comp:isInPackage [?node, ?package] :- 
    comp:isSelectedInternalOf[?node, ?package] .
    
We had no predicate for stating that something is a selected internal -
only that something was in the package. We need to be able to
differentiate the internals from the selected internals, hence the new
predicate comp:isSelectedInternalOf was introduced. The datalog rules is
extended to make it easier to handle the selectedInternals, when a node
is selected to be an internal node this also means that this node should
be in the commissioning package. The new datalog rule makes sure that
this happens automatically.

---------

Co-authored-by: Johannes Telle <johannes.telle@bouvet.no>
Co-authored-by: Dag Hovland <dhovl@equinor.com>
Co-authored-by: henriette.rebnor <henriette.rebnor@bouvet.no>
Co-authored-by: Embla Øye <118742425+eoye@users.noreply.github.com>