Trying out clickable area (#83)
## Aim of the PR
This PR fixes
[AB#208959](https://dev.azure.com/EquinorASA/bb9bd8cb-74f7-4ffa-b0cb-60eff0a0be58/_workitems/edit/208959).

Make it easier to click on components and lines

## Implementation 
Changed the StyledSvgElement and path that has the highlight to always
be rendered, just with opacity 0 when not selected. Then this larger
area is always clickable. Also moved the onClick in StyledPath to be in
the path (which has the highlight) for centerlines.

## Type of change
- [ ] Bug fix 
- [x] New feature 
- [ ] Breaking change 
- [ ] This change requires a documentation update

If the changes impact any dependent services then provide details.
## How Has This Been Tested?
Tried locally on Firefox in Ubuntu