Story 233061: Clean up frontend (#88)
## Aim of the PR
This PR fixes
[AB#233061](https://dev.azure.com/EquinorASA/bb9bd8cb-74f7-4ffa-b0cb-60eff0a0be58/_workitems/edit/233061)

This PR aims to clean up the frontend code a bit.

## Implementation 
**New context**: The old commissioning package context has been replaced
with a new context that uses a *reducer*. This is nice because a lot of
the time, we want to make updates both to the active package and to the
commissioning package list. The reducer is a function that takes in the
current commissioning package state, and an action, and returns the new
state. This way, instead of making sure we update the correct parts of
the context everywhere in the app, the reducer takes care of it for us.

**New context hook**: We have a new hook, useCommissioningPackages, that
returns the context and a dispatch. For read actions, we use the
context, and for write actions, we use the dispatch.

**CommissioningPackageActions**: These are functions that both call the
API and dispatch a reducer function. Use these in components when
possible.

**Helper functions**: Some new helper functions are created to avoid
duplicating so much code.

## Type of change
- [ ] Bug fix 
- [X] Refactor
- [ ] New feature 
- [ ] Breaking change 
- [ ] This change requires a documentation update

If the changes impact any dependent services then provide details.
## How Has This Been Tested?
Please describe the tests that you ran to verify that your changes works
as expected.

## Additional Changes
- Added TODOs in all files that can be removed when the new format and
API is complete.