Integration tests (#85)
## Aim of the PR
This PR fixes
[AB#223357](https://dev.azure.com/EquinorASA/bb9bd8cb-74f7-4ffa-b0cb-60eff0a0be58/_workitems/edit/223357)

Integration tests on the C# backend of the dexpi poc

## Implementation 
WebApplicationFactory for running the backend in tests. Dotnet
TestContainers for runing Rdfox. Some tests of package and boundaries.

## Type of change
- [ ] Bug fix 
- [x] New feature 
- [ ] Breaking change 
- [ ] This change requires a documentation update

If the changes impact any dependent services then provide details.
## How Has This Been Tested?
Ran the tests locally

## Additional Changes
The RdfoxApi was made non-static and included as a service. Upgraded to
dotnet 9.

---------

Co-authored-by: henriette.rebnor <henriette.rebnor@bouvet.no>
Co-authored-by: Johannes Telle <johannes.telle@bouvet.no>
Co-authored-by: Embla Øye <embla.oye@bouvet.no>