SVG Resizing (#97)
## Aim of the PR
This PR fixes
[AB#240703](https://dev.azure.com/EquinorASA/bb9bd8cb-74f7-4ffa-b0cb-60eff0a0be58/_workitems/edit/240703)

This PR aims to create a somewhat better user experience when
interacting with the SVG.

## Implementation 
Previously, the SVG container should have updated its size whenever the
user resized the window. This did not work as expected and led to some
jankiness.

I removed the SVG resizing, because it led to the zoom level resetting
every time you resized the window. Now, the SVG expands to fill the
available space on mount, and then you change the size by zooming in and
out.

The toolbar is also removed, but I kept the "minimap" for now.

## Type of change
- [X] Bug fix 
- [ ] New feature 
- [ ] Breaking change 
- [ ] This change requires a documentation update

If the changes impact any dependent services then provide details.
## How Has This Been Tested?
Please describe the tests that you ran to verify that your changes works
as expected.

## Additional Changes
- Removed unused packages

---------

Co-authored-by: Johannes Emil Telle <144658504+JohannesETelle@users.noreply.github.com>