Upgrade packages (#129)
[Pull requests practices for the SSI
team](https://github.com/equinor/ssi-infrastructure/blob/main/docs/pr_practices.md)

## Aim of the PR
This PR fixes
[AB#290714](https://dev.azure.com/EquinorASA/bb9bd8cb-74f7-4ffa-b0cb-60eff0a0be58/_workitems/edit/290714).

Upgrade packages.

## Implementation 
Upgraded all outdated packages in package.json. There was a mismatch in
some packages, for example `react` and `react-dom`, which resulted in
the demo not working. This is fixed now.

However, there is another error in the demo which is not related to the
package upgrades. I tried reverting back to before any major upgrades
were made, and we still got the same error. The error is related to our
triplestore responding with No Content when we expect a json object. I
have made a new User Story for this issue: [User Story
291301](https://dev.azure.com/EquinorASA/Spine/_workitems/edit/291301):
Dexpi demo is not able to get commissioning packages[User Story
291301](https://dev.azure.com/EquinorASA/Spine/_workitems/edit/291301):
Dexpi demo is not able to get commissioning packages

Upgrading these packages does not cause any more errors, so I think we
should merge this in so we at least have the latest updates, and then
try to figure out why it is failing in another User Story.

---------

Co-authored-by: daghovland <dhovl@equinor.com>